Pennsylvania Reemployment Bonus Experiment

Data from the Pennsylvania Reemployment Bonus experiment. The experiment randomly assigned unemployment insurance claimants to a treatment group that received a cash bonus for finding employment quickly.

data('Penn')

Format

A data.frame with 5099 observations and 23 variables, among them:

tg: =1 if assigned to the treatment group (reemployment bonus)

inuidur1: log unemployment duration

inuidur2: unemployment duration in weeks

female: =1 if female

black: =1 if Black

hispanic: =1 if Hispanic

othrace: =1 if other race

dep: number of dependents

q1-q6: quarter of entry indicators

recall: =1 if expecting recall

agelt35: =1 if age less than 35

agegt54: =1 if age greater than 54

durable: =1 if durable manufacturing industry

nondurable: =1 if nondurable manufacturing industry

lusd: =1 if low unemployment state duration

husd: =1 if high unemployment state duration

muld: =1 if mid unemployment state duration

Examples

data('Penn')
head(Penn)
#>     abdt tg inuidur1 inuidur2 female black hispanic othrace dep q1 q2 q3 q4 q5
#> 1  10824  0 2.890372       18      0     0        0       0   2  0  0  0  0  1
#> 4  10824  0 0.000000        1      0     0        0       0   0  0  0  0  0  1
#> 5  10747  0 3.295837       27      0     0        0       0   0  0  0  0  1  0
#> 12 10607  1 2.197225        9      0     0        0       0   0  0  0  1  0  0
#> 13 10831  0 3.295837       27      0     0        0       0   1  0  0  0  0  1
#> 14 10845  0 3.295837       27      1     0        0       0   0  0  0  0  0  1
#>    q6 recall agelt35 agegt54 durable nondurable lusd husd muld
#> 1   0      0       0       0       0          0    0    1    0
#> 4   0      0       0       0       0          0    1    0    0
#> 5   0      0       0       0       0          0    1    0    0
#> 12  0      0       1       0       0          0    0    0    1
#> 13  0      0       0       1       1          0    1    0    0
#> 14  0      0       0       1       0          0    1    0    0