Bounds on the omitted variable bias for causal machine learning

Computes confidence bounds on the target parameter of interest accounting for omitted variable biases.

dml_bounds(model, r2ya.dx, r2.rr, rho2 = 1)

confidence_bounds(model, ...)

# S3 method for numeric
confidence_bounds(
  theta.s,
  S2,
  se.theta.s,
  se.S2,
  cov.theta.S2,
  r2ya.dx,
  r2.rr,
  rho2 = 1,
  combine.method = "median",
  level = 0.95
)

# S3 method for dml
confidence_bounds(
  model,
  r2ya.dx,
  r2.rr,
  rho2 = 1,
  level = 0.95,
  combine.method = "median",
  ...
)

# S3 method for dml.bounds
confidence_bounds(
  model,
  r2ya.dx = NULL,
  r2.rr = NULL,
  rho2 = NULL,
  level = 0.95,
  combine.method = "median",
  return = c("lwr", "upr"),
  ...
)

Arguments

model

an object of class dml or dml.bounds.

r2ya.dx

(nonparametric) partial R2 of the omitted variables with the outcome. Must be a number between (0, 1).

r2.rr

how much variation latent variables create in the Riesz Representer of the target parameters. Must be a number between (0, 1). When the target of interest is the ATE in a partially linear model, this corresponds to the partial R2 of omitted variables with the treatment. When the target of interest is the ATE in a non-parametric model with a binary treatment, this corresponds to the gains in precision (i.e, 1/variance) when predicting who is assigned to treatment.

rho2

degree of adversity. Default is rho=1, which assumes the maximum degree of adversity of confounding.