Contour plots of omitted variable bias for Debiased Machine Learning

Contour plots of omitted variable bias for Debiased Machine Learning

ovb_contour_plot(model, ...)

# S3 method for dml
ovb_contour_plot(
  model,
  which.bound = c("lwr", "upr"),
  level = 0.95,
  combine.method = "median",
  rho2 = 1,
  r2ya.dx = NULL,
  r2.rr = r2ya.dx,
  bound.label = "Manual",
  group = F,
  group.number = 1,
  threshold = 0,
  lim.x = 0.15,
  lim.y = lim.x,
  asp = lim.x/lim.y,
  nlevels = 10,
  grid.number = 70,
  col.contour = "grey40",
  col.thr.line = "red",
  xlab = NULL,
  ylab = NULL,
  cex.lab = 0.8,
  cex.axis = 0.8,
  cex.main = 1,
  show.unadjusted = TRUE,
  label.unadjusted = "Unadjusted",
  label.text = TRUE,
  round = 0,
  cex.label.text = 0.7,
  label.bump.x = NULL,
  label.bump.y = NULL,
  list.par = NULL
)