coef, se, confint, print and summary methods for DML

The print and summary methods provide descriptions of the results obtained with the function dml.

The coef function extracts the coefficients.

The se function extracts the standard errors.

The confint function extracts the standard errors.

# S3 method for dml
summary(object, combine.method = "median", ...)

# S3 method for dml
coef(object, combine.method = "median", ...)

se(object, ...)

# S3 method for dml
se(object, combine.method = "median", ...)

# S3 method for dml
confint(object, params = NULL, level = 0.95, combine.method = "median", ...)

# S3 method for summary_dml
print(x, ...)

# S3 method for dml
print(
  x,
  digits = max(3L, getOption("digits") - 3L),
  combine.method = "median",
  ...
)

Arguments

object

an object of class dml.

combine.method

method to combine the results of each repetition of the DML fit. Options are mean and median. Default is median.

...

arguments passed to other methods.

params

character vector with the names of parameters.

level

confidence level. Default is 0.95.

x

an object of class dml.

digits

minimal number of significant digits.

Examples

# loads package
library(dml.sensemakr)

## loads data
data("pension")

# set the outcome
y <- pension$net_tfa  # net total financial assets

# set the treatment
d <- pension$e401    # 401K eligibility

# set the covariates (a matrix)
x <- model.matrix(~ -1 + age + inc  + educ+ fsize + marr + twoearn + pira + hown, data = pension)

## compute income quartiles for group ATE.
g1 <- cut(x[,"inc"], quantile(x[,"inc"], c(0, 0.25,.5,.75,1), na.rm = TRUE),
          labels = c("q1", "q2", "q3", "q4"), include.lowest = T)

# run DML (nonparametric model)
## 2 folds (change as needed)
## 1 repetition (change as needed)
dml.401k <- dml(y, d, x, model = "npm", groups = g1, cf.folds = 2, cf.reps = 1)
#> Debiased Machine Learning
#> 
#>  Model: Nonparametric 
#>  Target: ate 
#>  Cross-Fitting: 2 folds, 1 reps 
#>  ML Method: outcome (ranger), treatment (ranger)
#>  Tuning: dirty 
#> 
#> 
#> ====================================
#> Tuning parameters using all the data
#> ====================================
#> 
#> - Tuning Model for D.
#> -- Best Tune:
#>   mtry min.node.size splitrule
#> 1    2             5  variance
#> 
#> - Tuning Model for Y (non-parametric).
#> -- Best Tune:
#>   mtry min.node.size splitrule
#> 1    3             5  variance
#> 
#> 
#> ======================================
#> Repeating 2-fold cross-fitting 1 times
#> ======================================
#> 
#> -- Rep 1 -- Folds: 1  2  
#> 

summary(dml.401k)
#> 
#> Debiased Machine Learning
#> 
#>  Model: Nonparametric 
#>  Cross-Fitting: 2 folds, 1 reps 
#>  ML Method: outcome (ranger, R2 = 26.254%), treatment (ranger, R2 = 10.865%)
#>  Tuning: dirty 
#> 
#> Average Treatment Effect: 
#> 
#>     Estimate Std. Error t value   P(>|t|)    
#> ate   7738.1     1174.5  6.5881 4.454e-11 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Group Average Treatment Effect: 
#> 
#>         Estimate Std. Error t value   P(>|t|)    
#> gate.q1  5126.07     890.39  5.7571 8.556e-09 ***
#> gate.q2  3171.44    1163.51  2.7257  0.006416 ** 
#> gate.q3  6101.92    1966.35  3.1032  0.001915 ** 
#> gate.q4 16550.60    4001.98  4.1356 3.540e-05 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Note: DML estimates combined using the median method.
summary(dml.401k, combine.method = "mean")
#> 
#> Debiased Machine Learning
#> 
#>  Model: Nonparametric 
#>  Cross-Fitting: 2 folds, 1 reps 
#>  ML Method: outcome (ranger, R2 = 26.254%), treatment (ranger, R2 = 10.865%)
#>  Tuning: dirty 
#> 
#> Average Treatment Effect: 
#> 
#>     Estimate Std. Error t value   P(>|t|)    
#> ate   7738.1     1174.5  6.5881 4.454e-11 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Group Average Treatment Effect: 
#> 
#>         Estimate Std. Error t value   P(>|t|)    
#> gate.q1  5126.07     890.39  5.7571 8.556e-09 ***
#> gate.q2  3171.44    1163.51  2.7257  0.006416 ** 
#> gate.q3  6101.92    1966.35  3.1032  0.001915 ** 
#> gate.q4 16550.60    4001.98  4.1356 3.540e-05 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Note: DML estimates combined using the mean method.
coef(dml.401k)
#>       ate   gate.q1   gate.q2   gate.q3   gate.q4 
#>  7738.065  5126.065  3171.442  6101.917 16550.601 
coef(dml.401k, combine.method = "mean")
#>       ate   gate.q1   gate.q2   gate.q3   gate.q4 
#>  7738.065  5126.065  3171.442  6101.917 16550.601 
se(dml.401k)
#>       ate   gate.q1   gate.q2   gate.q3   gate.q4 
#> 1174.5456  890.3865 1163.5136 1966.3522 4001.9779 
confint(dml.401k, combine.method = "mean")
#>            2.5 %    97.5 %
#> ate     5435.999 10040.132
#> gate.q1 3380.940  6871.191
#> gate.q2  890.997  5451.887
#> gate.q3 2247.938  9955.897
#> gate.q4 8706.868 24394.333